Fix some link errors and a few remaining compile warnings (the ones that aren't bugs)
- if static const int foo=VALUE; is in header, don't need to have
  a definition in a .cc file.  The definition goes into every .cc
  file with "vague linkage"
- add sqlite and lua to the sln/vcproj



git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@3576 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/luadgn.cc
@@ -1485,7 +1485,7 @@ static int dgn_random_walk(lua_State *ls)
     // Allow movement to all 8 adjacent squares if distance is 1
     // (needed since diagonal moves are distance sqrt(2))
     if (dist == 1)
-        dist_left = SQRT_2;
+        dist_left = (float)SQRT_2;
 
     int moves_left = dist;
     coord_def pos(x, y);
